# openapi.yaml
...
openapi: 3.1.0
info:
  title: photo-portfolio-store
  version: 0.1.0
  description: |
    Photographer portfolio + print storefront API.

    Bounded contexts: auth, events, media, search, catalog, commerce,
    compliance, admin. Each operation declares its context via tag.
servers:
  - url: http://localhost:4000
    description: local
...
security: []
paths:
...
  /v1/consents/biometric:
    post:
      operationId: grantBiometricConsent
      tags: [compliance]
      summary: Grant biometric consent for face search in an event
      x-status: stub
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConsentGrantRequest'
      responses:
        '201':
          description: Consent recorded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Consent'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/Unprocessable' }
        '500': { $ref: '#/components/responses/ServerError' }
  /v1/consents/biometric/{id}:
    parameters:
      - $ref: '#/components/parameters/IdParam'
    delete:
      operationId: revokeBiometricConsent
      tags: [compliance]
      summary: Revoke a biometric consent; purges derived data
      x-status: stub
      security: [{ bearerAuth: [] }]
      responses:
        '204': { description: Consent revoked. }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/ServerError' }
  /v1/me/biometric-data:
    get:
      operationId: getMyBiometricData
      tags: [compliance]
      summary: Export the caller's biometric records
      x-status: stub
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Biometric records and consents.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BiometricExport'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '500': { $ref: '#/components/responses/ServerError' }
...
  /v1/takedowns:
    post:
      operationId: submitTakedown
      tags: [compliance]
      summary: Submit a takedown request (anonymous; email-verified)
      description: Anonymous-allowed, rate-limited 5 per IP per hour. Creates a request and emails the subject a verification link.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [subjectEmail, reason, legalBasis]
              properties:
                subjectEmail: { type: string, format: email, maxLength: 320 }
                reason: { type: string, enum: [lgpd, gdpr, bipa, copyright, other] }
                legalBasis: { type: string, minLength: 1, maxLength: 2000 }
                photoIds:
                  type: array
                  items: { type: string, format: uuid }
                  maxItems: 500
                evidenceUrl: { type: string, format: uri, maxLength: 1000 }
                notes: { type: string, maxLength: 2000 }
      responses:
        '202':
          description: Request accepted; verification email sent.
          content:
            application/json:
              schema:
                type: object
                required: [trackingId]
                properties:
                  trackingId: { type: string, format: uuid }
        '400': { $ref: '#/components/responses/BadRequest' }
        '500': { $ref: '#/components/responses/ServerError' }
  /v1/takedowns/{id}/verify:
    parameters:
      - $ref: '#/components/parameters/IdParam'
    get:
      operationId: verifyTakedown
      tags: [compliance]
      summary: Consume the email-verification token for a takedown request
      parameters:
        - name: token
          in: query
          required: true
          schema: { type: string, minLength: 8, maxLength: 256 }
      responses:
        '200':
          description: Verified; request advanced to the admin queue.
          content:
            application/json:
              schema:
                type: object
                required: [status]
                properties:
                  status: { type: string, enum: [verifying] }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '410':
          description: Token expired.
        '500': { $ref: '#/components/responses/ServerError' }
  /v1/takedowns/{id}:
    parameters:
      - $ref: '#/components/parameters/IdParam'
    get:
      operationId: getTakedownStatus
      tags: [compliance]
      summary: Token-gated subject view of a takedown request status
      parameters:
        - name: token
          in: query
          required: true
          schema: { type: string, minLength: 8, maxLength: 256 }
      responses:
        '200':
          description: Takedown status.
          content:
            application/json:
              schema:
                type: object
                required: [status, receivedAt, slaDueAt, fulfilledAt]
                properties:
                  status: { type: string }
                  receivedAt: { type: string, format: date-time }
                  slaDueAt: { type: string, format: date-time }
                  fulfilledAt: { type: string, format: date-time, nullable: true }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/ServerError' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdParam:
      in: path
      name: id
      required: true
      schema:
        $ref: '#/components/schemas/Uuid'
...
  responses:
    BadRequest:
      description: Malformed request.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Unauthorized:
      description: Missing or invalid credentials.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Forbidden:
      description: Authenticated but not permitted.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Conflict:
      description: Conflict with current resource state.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Unprocessable:
      description: Validation failed.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    ServerError:
      description: Unexpected server error.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
  schemas:
    Uuid:
      type: string
      format: uuid
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Stable machine-readable error code.
            message:
              type: string
            details:
              type: object
              additionalProperties: true
            trace_id:
              type: string
...
    ConsentScope:
      type: string
      enum: [biometric, marketing, terms_of_service, privacy_policy]
    ConsentJurisdiction:
      type: string
      enum: [eu_gdpr, br_lgpd, us_bipa, us_ccpa, other]
    Consent:
      type: object
      required: [id, scope, grantedAt, jurisdiction, consentVersion, createdAt]
      properties:
        id: { $ref: '#/components/schemas/Uuid' }
        scope: { $ref: '#/components/schemas/ConsentScope' }
        subjectId: { $ref: '#/components/schemas/Uuid' }
        eventId: { $ref: '#/components/schemas/Uuid' }
        grantedAt: { type: string, format: date-time }
        revokedAt: { type: string, format: date-time, nullable: true }
        retentionUntil: { type: string, format: date-time, nullable: true }
        jurisdiction: { $ref: '#/components/schemas/ConsentJurisdiction' }
        consentVersion: { type: string }
        createdAt: { type: string, format: date-time }
    ConsentGrantRequest:
      type: object
      required: [eventId, jurisdiction, consentVersion]
      properties:
        eventId: { $ref: '#/components/schemas/Uuid' }
        jurisdiction: { $ref: '#/components/schemas/ConsentJurisdiction' }
        consentVersion: { type: string }
        locale: { type: string }
    BiometricExport:
      type: object
      required: [consents, faceVectorCount]
      properties:
        consents:
          type: array
          items: { $ref: '#/components/schemas/Consent' }
        faceVectorCount: { type: integer }
...